refactor(codex): dedupe stripUserMessagePrefix + stripLeadingCodexContextBlocks
Two pure helpers were defined byte-for-byte identically in both
codex/discover.ts and codex/parser.ts. Move them to the existing
codex/constants.ts (which already houses pure helpers like
isCodexToolCallType) and import from there.

No behavior change — the helpers are pure functions of their inputs.
