+++ packages/cli/src/providers/codex/parser.ts
@@ -4,7 +4,11 @@ import { extname } from "node:path";
 import { estimateActiveDuration } from "../../duration.js";
 import type { ContentBlock, ParsedTurn, SessionInfo } from "../../types.js";
 import type { Compaction, ProviderParseResult, TokenUsage } from "../types.js";
-import { CODEX_CONTEXT_TAGS, isCodexToolCallType, USER_MESSAGE_BEGIN } from "./constants.js";
+import {
+  isCodexToolCallType,
+  stripLeadingCodexContextBlocks,
+  stripUserMessagePrefix,
+} from "./constants.js";
 
 interface PendingTool {
   id: string;
@@ -666,30 +670,6 @@ function normalizeUserMessageText(text: string): string {
     normalized = stripUserMessagePrefix(stripLeadingCodexContextBlocks(normalized)).trim();
   }
   return normalized;
-}
-
-function stripLeadingCodexContextBlocks(text: string): string {
-  let remaining = text.trim();
-  let stripped = true;
-  while (stripped) {
-    stripped = false;
-    for (const tag of CODEX_CONTEXT_TAGS) {
-      const open = `<${tag}>`;
-      const close = `</${tag}>`;
-      if (!remaining.startsWith(open)) continue;
-      const closeIndex = remaining.indexOf(close);
-      if (closeIndex === -1) return "";
-      remaining = remaining.slice(closeIndex + close.length).trim();
-      stripped = true;
-      break;
-    }
-  }
-  return remaining;
-}
-
-function stripUserMessagePrefix(text: string): string {
-  const idx = text.indexOf(USER_MESSAGE_BEGIN);
-  return idx === -1 ? text : text.slice(idx + USER_MESSAGE_BEGIN.length);
 }
 
 function localImageToDataUrl(path: string): string | undefined {
